:fu: Implement Helpers module
